# automacao/barrigarest-openapi.yaml
openapi: 3.0.3
info:
  title: Barriga API
  version: 1.0.0
  description: Endpoints para autenticação e gerenciamento de contas na API Barriga.
servers:
  - url: https://barrigarest.wcaquino.me/
paths:
  /signin:
    post:
      summary: Autenticação do usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: a@a
                senha:
                  type: string
                  example: a
      responses:
        '200':
          description: Token JWT
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
  /contas:
    post:
      summary: Criação de conta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                  example: Conta Exemplo
      responses:
        '201':
          description: Conta criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  nome:
                    type: string
        '400':
          description: Conta já existente
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    get:
      summary: Listar contas (necessita autenticação)
      responses:
        '200':
          description: Lista de contas
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    nome:
                      type: string
  /contas/{id}:
    put:
      summary: Atualizar conta
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
      responses:
        '200':
          description: Conta atualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  nome:
                    type: string
    delete:
      summary: Remover conta
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Conta removida com sucesso